Get andoid nearly working
